refactor: Refactor code 5
